pyphotobooth/requirements.txt
picamera>=1.13
pillow>=8.3.2
pil
PyYAML>=5.4
imutils>=0.5
numpy>=1.21
opencv-python>=4.5.3
